Renamed server/auth/index.js into server/auth/auth.js
